@@ -10,7 +10,7 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "16"
+    "node": "20"
   },
   "main": "lib/index.js",
   "dependencies": {
@@ -20,9 +20,9 @@
     "cors": "^2.8.5",
     "date-fns": "^2.29.3",
     "esm": "^3.2.25",
-    "express": "^4.18.2",
-    "firebase-admin": "^11.5.0",
-    "firebase-functions": "^4.3.0",
+  "express": "^4.18.2",
+  "firebase-admin": "^11.10.0",
+  "firebase-functions": "^5.1.0",
     "gunzip": "^1.2.2",
     "gunzip-file": "^0.1.1",
     "p-retry": "^5.1.2",
