@@ -3,9 +3,21 @@
  * Ubicación: src/app/app.component.ts
  */
 
-import { Component, OnInit, OnDestroy, Inject, PLATFORM_ID, inject } from '@angular/core';
+import {
+  Component,
+  OnInit,
+  OnDestroy,
+  Inject,
+  PLATFORM_ID,
+  inject,
+} from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { Router, RouterOutlet, NavigationEnd, NavigationStart } from '@angular/router';
+import {
+  Router,
+  RouterOutlet,
+  NavigationEnd,
+  NavigationStart,
+} from '@angular/router';
 import { filter, takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
 
@@ -15,6 +27,7 @@ import { RightSidebarComponent } from './components/right-sidebar/right-sidebar.
 
 // Servicios SOLID (opcional - solo para logging mejorado)
 import { ConsoleLoggerService } from './services/core/logger.service';
+import { FooterComponent } from './components/footer/footer.component';
 
 @Component({
   selector: 'app-root',
@@ -23,10 +36,11 @@ import { ConsoleLoggerService } from './services/core/logger.service';
     CommonModule,
     RouterOutlet,
     LeftSidebarComponent,
-    RightSidebarComponent
+    RightSidebarComponent,
+    FooterComponent,
   ],
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent implements OnInit, OnDestroy {
   private static instanceCount = 0;
@@ -34,69 +48,88 @@ export class AppComponent implements OnInit, OnDestroy {
   private destroy$ = new Subject<void>();
   private routerEventsDisabled = false;
   private navigationCount = 0;
-  
+
   // Integración opcional con SOLID logging
   private logger = inject(ConsoleLoggerService, { optional: true });
-  
+
   constructor(
     @Inject(PLATFORM_ID) private readonly platformId: Object,
     private readonly router: Router
   ) {
     AppComponent.instanceCount++;
     this.instanceId = `app-${AppComponent.instanceCount}`;
-    
+
     // Logging mejorado con SOLID (fallback a console si no está disponible)
     this.logInfo(`🏗️ APP COMPONENT - Instancia ${this.instanceId} creada`);
-    
+
     // Protección contra múltiples instancias del AppComponent
     if (AppComponent.instanceCount > 1) {
-      this.logWarning(`⚠️ MÚLTIPLES INSTANCIAS DE APP COMPONENT DETECTADAS: ${AppComponent.instanceCount}`);
-      this.logWarning('🔄 Instancia adicional iniciada (puede ser normal en desarrollo con HMR)');
-      
+      this.logWarning(
+        `⚠️ MÚLTIPLES INSTANCIAS DE APP COMPONENT DETECTADAS: ${AppComponent.instanceCount}`
+      );
+      this.logWarning(
+        '🔄 Instancia adicional iniciada (puede ser normal en desarrollo con HMR)'
+      );
+
       // En lugar de return, continuamos pero con logging mejorado
     }
-    
+
     // Monitorear navegación con protección mejorada
-    this.router.events.pipe(
-      filter(event => event instanceof NavigationStart || event instanceof NavigationEnd),
-      takeUntil(this.destroy$)
-    ).subscribe(event => {
-      if (this.routerEventsDisabled) {
-        this.logWarning('🚫 Eventos de router deshabilitados para evitar bucle');
-        return;
-      }
-      
-      if (event instanceof NavigationStart) {
-        this.navigationCount++;
-        this.logDebug(`🧭 Navegación iniciada: ${event.url} (count: ${this.navigationCount})`);
-        
-        // Protección mejorada contra bucles de navegación
-        if (this.navigationCount > 5) {
-          this.logError('⚠️ BUCLE DE NAVEGACIÓN DETECTADO - DESHABILITANDO EVENTOS');
-          this.logError('🛑 ROUTER EVENTS DISABLED PARA PREVENIR BUCLE INFINITO');
-          this.routerEventsDisabled = true;
-          this.destroy$.next();
+    this.router.events
+      .pipe(
+        filter(
+          (event) =>
+            event instanceof NavigationStart || event instanceof NavigationEnd
+        ),
+        takeUntil(this.destroy$)
+      )
+      .subscribe((event) => {
+        if (this.routerEventsDisabled) {
+          this.logWarning(
+            '🚫 Eventos de router deshabilitados para evitar bucle'
+          );
           return;
         }
-      } else if (event instanceof NavigationEnd) {
-        this.logDebug(`✅ Navegación completada: ${event.url}`);
-        
-        // Resetear contador después de navegación exitosa
-        setTimeout(() => {
-          this.navigationCount = 0;
-          this.logDebug('🔄 Contador de navegación reseteado');
-        }, 2000);
-      }
-    });
+
+        if (event instanceof NavigationStart) {
+          this.navigationCount++;
+          this.logDebug(
+            `🧭 Navegación iniciada: ${event.url} (count: ${this.navigationCount})`
+          );
+
+          // Protección mejorada contra bucles de navegación
+          if (this.navigationCount > 5) {
+            this.logError(
+              '⚠️ BUCLE DE NAVEGACIÓN DETECTADO - DESHABILITANDO EVENTOS'
+            );
+            this.logError(
+              '🛑 ROUTER EVENTS DISABLED PARA PREVENIR BUCLE INFINITO'
+            );
+            this.routerEventsDisabled = true;
+            this.destroy$.next();
+            return;
+          }
+        } else if (event instanceof NavigationEnd) {
+          this.logDebug(`✅ Navegación completada: ${event.url}`);
+
+          // Resetear contador después de navegación exitosa
+          setTimeout(() => {
+            this.navigationCount = 0;
+            this.logDebug('🔄 Contador de navegación reseteado');
+          }, 2000);
+        }
+      });
   }
 
   ngOnInit(): void {
     this.logInfo(`🚀 APP COMPONENT INIT - Instancia ${this.instanceId}`);
     this.logInfo('✅ Arquitectura SOLID inicializada');
-    
+
     // Protección adicional en ngOnInit
     if (AppComponent.instanceCount > 1) {
-      this.logWarning(`⚠️ Múltiples instancias detectadas en ngOnInit (${AppComponent.instanceCount})`);
+      this.logWarning(
+        `⚠️ Múltiples instancias detectadas en ngOnInit (${AppComponent.instanceCount})`
+      );
       // Continuar con la inicialización normal
     }
 
@@ -105,20 +138,22 @@ export class AppComponent implements OnInit, OnDestroy {
       instanceId: this.instanceId,
       instanceCount: AppComponent.instanceCount,
       platformId: this.platformId,
-      routerEventsDisabled: this.routerEventsDisabled
+      routerEventsDisabled: this.routerEventsDisabled,
     });
   }
 
   ngOnDestroy(): void {
     this.logInfo(`🗑️ APP COMPONENT DESTROY - Instancia ${this.instanceId}`);
-    
+
     this.destroy$.next();
     this.destroy$.complete();
-    
+
     // Decrementar contador de instancias
     AppComponent.instanceCount = Math.max(0, AppComponent.instanceCount - 1);
-    
-    this.logDebug(`📉 Instance count after destroy: ${AppComponent.instanceCount}`);
+
+    this.logDebug(
+      `📉 Instance count after destroy: ${AppComponent.instanceCount}`
+    );
   }
 
   // ===============================================
@@ -170,7 +205,7 @@ export class AppComponent implements OnInit, OnDestroy {
       instanceCount: AppComponent.instanceCount,
       routerEventsDisabled: this.routerEventsDisabled,
       navigationCount: this.navigationCount,
-      isDestroyed: this.destroy$.closed
+      isDestroyed: this.destroy$.closed,
     };
 
     this.logDebug('=== APP COMPONENT DEBUG STATE ===');
