@@ -1,13 +1,16 @@
 import { mergeApplicationConfig, ApplicationConfig } from '@angular/core';
 import { provideServerRendering } from '@angular/platform-server';
 import { provideClientHydration } from '@angular/platform-browser';
+import { provideNoopAnimations } from '@angular/platform-browser/animations';
 import { appConfig } from './app.config';
 
 const serverConfig: ApplicationConfig = {
   providers: [
     provideServerRendering(),
+    // Desactivar animaciones en el servidor para evitar errores al procesar triggers
+    provideNoopAnimations(),
     provideClientHydration(),
-  ]
+  ],
 };
 
 export const config = mergeApplicationConfig(appConfig, serverConfig);
