@@ -1,14 +1,25 @@
 import { CommonModule } from '@angular/common';
-import { Component, Input, OnInit, OnChanges, SimpleChanges } from '@angular/core';
+import {
+  Component,
+  Input,
+  OnInit,
+  OnChanges,
+  SimpleChanges,
+} from '@angular/core';
 import { Router } from '@angular/router';
+import { slugify } from 'src/app/utils/utils';
 import { TvGuideService } from 'src/app/services/tv-guide.service';
-import { diffHour, getHoraInicio, formatCorrectTime } from 'src/app/utils/utils';
-import { 
-  IBannerData, 
-  IBannerInputData, 
-  IImageService, 
+import {
+  diffHour,
+  getHoraInicio,
+  formatCorrectTime,
+} from 'src/app/utils/utils';
+import {
+  IBannerData,
+  IBannerInputData,
+  IImageService,
   IBannerDataService,
-  ITimeUtilsService 
+  ITimeUtilsService,
 } from '../../interfaces/banner.interface';
 
 @Component({
@@ -18,9 +29,22 @@ import {
   standalone: true,
   imports: [CommonModule],
 })
-export class BannerComponent implements OnInit, OnChanges, IImageService, IBannerDataService, ITimeUtilsService {
+export class BannerComponent
+  implements
+    OnInit,
+    OnChanges,
+    IImageService,
+    IBannerDataService,
+    ITimeUtilsService
+{
   @Input() data: IBannerInputData = {};
-  
+  /** If true, the banner adapts to small containers (fills parent) */
+  @Input() compact: boolean = false;
+  /** If true, use a small square channel logo instead of the large rectangular one */
+  @Input() compactLogo: boolean = false;
+  /** If true, hide the top-right time display */
+  @Input() hideTopTime: boolean = false;
+
   public bannerData: IBannerData | null = null;
   public logo: string = '';
   public time: string = '';
@@ -28,7 +52,6 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
   constructor(private router: Router, private guiatvSvc: TvGuideService) {}
 
   ngOnInit(): void {
-    console.log('BannerComponent initialized with data:', this.data);
     this.processBannerData();
   }
 
@@ -38,16 +61,12 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
     }
   }
 
-  /**
-   * Procesa los datos de entrada y los convierte al formato del banner
-   */
   private processBannerData(): void {
     if (!this.data || Object.keys(this.data).length === 0) {
       this.bannerData = null;
       return;
     }
 
-    // Detectar el tipo de datos y convertir apropiadamente
     if (this.isMovieData(this.data)) {
       this.bannerData = this.convertMovieToBannerData(this.data);
     } else if (this.isProgramData(this.data)) {
@@ -56,60 +75,64 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
       this.bannerData = this.convertGenericToBannerData(this.data);
     }
 
-    // Calcular tiempo después de la conversión
     if (this.bannerData?.start && this.bannerData?.stop) {
-      this.time = this.calculateDuration(this.bannerData.start, this.bannerData.stop);
+      this.time = this.calculateDuration(
+        this.bannerData.start,
+        this.bannerData.stop
+      );
     }
   }
-  /**
-   * Detecta si los datos son de una película destacada
-   */
+
   private isMovieData(data: IBannerInputData): boolean {
-    return !!(data.title && (data.poster || data.description || data.rating || data.releaseDate));
+    return !!(
+      data.title &&
+      (data.poster || data.description || data.rating || data.releaseDate)
+    );
   }
 
-  /**
-   * Detecta si los datos son de un programa de TV
-   */
   private isProgramData(data: IBannerInputData): boolean {
     return !!(data.title && data.channel && data.start && data.stop);
   }
 
-  // Implementación de IBannerDataService
   convertMovieToBannerData(movieData: IBannerInputData): IBannerData {
-    const title = typeof movieData.title === 'string' 
-      ? { value: movieData.title } 
-      : movieData.title || { value: 'Título desconocido' };
-
-    const descData = typeof movieData.desc === 'string' 
-      ? { details: movieData.desc } 
-      : movieData.desc;
+    const title =
+      typeof movieData.title === 'string'
+        ? { value: movieData.title }
+        : movieData.title || { value: 'Título desconocido' };
+    const descData =
+      typeof movieData.desc === 'string'
+        ? { details: movieData.desc }
+        : movieData.desc;
 
     return {
       title,
-      channel: movieData.channelName || movieData.channel || 'Canal desconocido',
+      channel:
+        movieData.channelName || movieData.channel || 'Canal desconocido',
       poster: movieData.poster || movieData.icon,
       icon: movieData.icon,
       start: this.normalizeTimeString(movieData.startTime || movieData.start),
       stop: this.normalizeTimeString(movieData.endTime || movieData.stop),
       desc: {
-        details: movieData.description || descData?.details || movieData.overview,
+        details:
+          movieData.description || descData?.details || movieData.overview,
         year: movieData.year || movieData.releaseDate || descData?.year,
-        rate: movieData.rating?.toString() || descData?.rate || 'TP'
+        rate: movieData.rating?.toString() || descData?.rate || 'TP',
       },
-      category: typeof movieData.category === 'string' 
-        ? { value: movieData.category } 
-        : movieData.category,
-      starRating: movieData.starRating || movieData.vote_average || movieData.rating,
-      id: movieData.id
+      category:
+        typeof movieData.category === 'string'
+          ? { value: movieData.category }
+          : movieData.category,
+      starRating:
+        movieData.starRating || movieData.vote_average || movieData.rating,
+      id: movieData.id,
     };
   }
 
   convertProgramToBannerData(programData: IBannerInputData): IBannerData {
-    const title = typeof programData.title === 'string' 
-      ? { value: programData.title } 
-      : programData.title || { value: 'Programa desconocido' };
-
+    const title =
+      typeof programData.title === 'string'
+        ? { value: programData.title }
+        : programData.title || { value: 'Programa desconocido' };
     return {
       title,
       channel: programData.channel || 'Canal desconocido',
@@ -117,29 +140,28 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
       icon: programData.icon,
       start: this.normalizeTimeString(programData.start),
       stop: this.normalizeTimeString(programData.stop),
-      desc: typeof programData.desc === 'string' 
-        ? { details: programData.desc } 
-        : programData.desc,
-      category: typeof programData.category === 'string' 
-        ? { value: programData.category } 
-        : programData.category,
+      desc:
+        typeof programData.desc === 'string'
+          ? { details: programData.desc }
+          : programData.desc,
+      category:
+        typeof programData.category === 'string'
+          ? { value: programData.category }
+          : programData.category,
       starRating: programData.starRating,
-      id: programData.id
+      id: programData.id,
     };
   }
 
   convertGenericToBannerData(data: IBannerInputData): IBannerData {
-    // Fallback genérico
     return this.convertProgramToBannerData(data);
   }
 
-  // Implementación de ITimeUtilsService
   formatTime(timeString: string): string {
     try {
-      // Usar la utilidad centralizada para consistencia
       return formatCorrectTime(timeString);
     } catch (error) {
-      return getHoraInicio(timeString); // Fallback a la utilidad original
+      return getHoraInicio(timeString);
     }
   }
 
@@ -153,13 +175,48 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
     return time;
   }
 
-  // Implementación de IImageService
   getChannelLogoUrl(channelName: string): string {
     return `https://wsrv.nl/?url=https://raw.githubusercontent.com/davidmuma/picons_dobleM/master/icon/${channelName}.png`;
   }
 
   getProgramPosterUrl(programData: any): string {
-    return programData?.poster || programData?.icon || this.getFallbackImageUrl();
+    return (
+      programData?.poster || programData?.icon || this.getFallbackImageUrl()
+    );
+  }
+
+  // Template helper: wrapper for encodeURIComponent so the template can call it
+  encodeURIComponent(v: any): string {
+    try {
+      return globalThis.encodeURIComponent(String(v || ''));
+    } catch (_) {
+      return String(v || '');
+    }
+  }
+
+  // Build srcset for the large banner background (desktop sizes)
+  getBannerSrcset(raw: string): string {
+    if (!raw) return '';
+    const sizes = [768, 1024, 1280, 1600, 1920];
+    return sizes
+      .map(
+        (w) =>
+          `https://wsrv.nl/?url=${this.encodeURIComponent(
+            raw
+          )}&w=${w}&h=${Math.round(w * 0.416)}&output=webp ${w}w`
+      )
+      .join(', ');
+  }
+
+  // Stub for reminder action — keep minimal to avoid runtime errors; can be extended later
+  addReminder(): void {
+    // For now simply log and optionally call a service later
+    try {
+      console.log(
+        'addReminder clicked for',
+        this.bannerData?.id || this.bannerData?.title?.value
+      );
+    } catch (_) {}
   }
 
   getFallbackImageUrl(): string {
@@ -171,13 +228,8 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
     target.src = this.getFallbackImageUrl();
   }
 
-  // Métodos públicos para el template
   public getHora(hora: string): string {
     return this.formatTime(hora);
-  }
-
-  private getTime(start: string, stop: string): string {
-    return this.calculateDuration(start, stop);
   }
 
   public onImageError(event: Event): void {
@@ -186,13 +238,20 @@ export class BannerComponent implements OnInit, OnChanges, IImageService, IBanne
 
   public navigateTo(): void {
     if (!this.bannerData) return;
-    
-    this.guiatvSvc.setDetallesPrograma(this.data);
+
+    // Use the normalized bannerData so the details component gets a consistent shape
+    this.guiatvSvc.setDetallesPrograma(this.bannerData);
 
     const titleValue = this.bannerData.title?.value || 'unknown';
-    this.router.navigate([
-      'programacion-tv/detalles',
-      titleValue.replace(/\s/g, '-'),
-    ]);
+    const slug = slugify(titleValue);
+    if (this.isMovieData(this.data)) {
+      this.router.navigate(['/peliculas', slug], {
+        state: { bannerData: this.bannerData },
+      });
+    } else {
+      this.router.navigate(['/programas', slug], {
+        state: { bannerData: this.bannerData },
+      });
+    }
   }
 }
