@@ -7,10 +7,12 @@ import { CommonModule } from '@angular/common';
 import { Component, OnInit, inject, DestroyRef, signal } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Router } from '@angular/router';
+import { slugify } from 'src/app/utils/utils';
 import { AutocompleteComponent } from '../autocomplete/autocomplete.component';
 
 // NUEVO: Usar HomeDataService en lugar de TvGuideService
 import { HomeDataService } from 'src/app/services/features/home-data.service';
+import { TvGuideService } from 'src/app/services/tv-guide.service';
 import { IFeaturedMovie } from 'src/app/interfaces';
 import { environment } from 'src/environments/environment';
 
@@ -19,7 +21,7 @@ import { environment } from 'src/environments/environment';
 // ===============================================
 
 interface ISidebarItem {
-  id?: string;
+  id?: string | number;
   title: { value: string };
   category: { value: string };
   icon: string;
@@ -34,19 +36,19 @@ interface ISidebarItem {
   imports: [CommonModule, AutocompleteComponent],
 })
 export class RightSidebarComponent implements OnInit {
-  
   // ===============================================
   // DEPENDENCY INJECTION
   // ===============================================
-  
+
   private readonly homeDataService = inject(HomeDataService);
+  private readonly tvGuide = inject(TvGuideService);
   private readonly router = inject(Router);
   private readonly destroyRef = inject(DestroyRef);
 
   // ===============================================
   // COMPONENT STATE - SIGNALS CORREGIDOS
   // ===============================================
-  
+
   public readonly popular_movies = signal<ISidebarItem[]>([]);
   public readonly popular_series = signal<ISidebarItem[]>([]);
   public readonly isLoading = signal<boolean>(true);
@@ -58,6 +60,19 @@ export class RightSidebarComponent implements OnInit {
   ngOnInit(): void {
     console.log('🔧 RIGHT-SIDEBAR - Initializing with corrected SOLID system');
     this.initializeDataStreams();
+    // If the featured lists are empty (app didn't start on Home), trigger load
+    try {
+      if (this.popular_movies().length === 0) {
+        console.log('🔄 RIGHT-SIDEBAR - Triggering peliculas destacadas load');
+        this.tvGuide.setPeliculasDestacadas();
+      }
+      if (this.popular_series().length === 0) {
+        console.log('🔄 RIGHT-SIDEBAR - Triggering series destacadas load');
+        this.tvGuide.setSeriesDestacadas();
+      }
+    } catch (e) {
+      // ignore
+    }
   }
 
   // ===============================================
@@ -73,43 +88,56 @@ export class RightSidebarComponent implements OnInit {
     // Stream de películas populares
     this.homeDataService.popularMovies$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(movies => {
-        console.log('🎬 RIGHT-SIDEBAR - Popular movies received:', movies.length);
-        
+      .subscribe((movies) => {
+        console.log(
+          '🎬 RIGHT-SIDEBAR - Popular movies received:',
+          movies.length
+        );
+
         if (movies && movies.length > 0) {
           // Convertir al formato que espera el template
           const convertedMovies = this.convertToSidebarFormat(movies, 'movie');
           this.popular_movies.set(convertedMovies);
           this.isLoading.set(false);
-          
-          console.log('✅ RIGHT-SIDEBAR - Movies converted and set:', convertedMovies.length);
+
+          console.log(
+            '✅ RIGHT-SIDEBAR - Movies converted and set:',
+            convertedMovies.length
+          );
         } else {
-          console.log('⚠️ RIGHT-SIDEBAR - No movies received, keeping loading state');
+          console.log(
+            '⚠️ RIGHT-SIDEBAR - No movies received, keeping loading state'
+          );
         }
       });
 
     // Stream de series (simuladas desde películas por ahora)
-    this.homeDataService.popularMovies$
+    this.tvGuide
+      .getSeriesDestacadas()
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(movies => {
-        if (movies && movies.length > 0) {
-          // Simular series para mantener funcionalidad
-          const convertedSeries = this.convertToSidebarFormat(movies, 'series');
+      .subscribe((series) => {
+        if (series && series.length > 0) {
+          const convertedSeries = this.convertToSidebarFormat(series, 'series');
           this.popular_series.set(convertedSeries);
-          
-          console.log('📺 RIGHT-SIDEBAR - Series simulated and set:', convertedSeries.length);
+
+          console.log(
+            '📺 RIGHT-SIDEBAR - Series received and set:',
+            convertedSeries.length
+          );
         }
       });
 
     // Stream de estado de carga
     this.homeDataService.loading$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(loading => {
+      .subscribe((loading) => {
         if (!loading && this.popular_movies().length === 0) {
           // Si ya no está cargando pero no hay películas, mantener estado de carga local
           setTimeout(() => {
             if (this.popular_movies().length === 0) {
-              console.log('⏰ RIGHT-SIDEBAR - Loading timeout, no movies available');
+              console.log(
+                '⏰ RIGHT-SIDEBAR - Loading timeout, no movies available'
+              );
               this.isLoading.set(false);
             }
           }, 2000);
@@ -124,22 +152,35 @@ export class RightSidebarComponent implements OnInit {
   /**
    * Convierte IFeaturedMovie al formato que espera el template del sidebar
    */
-  private convertToSidebarFormat(movies: IFeaturedMovie[], type: 'movie' | 'series'): ISidebarItem[] {
+  private convertToSidebarFormat(
+    movies: any[],
+    type: 'movie' | 'series'
+  ): ISidebarItem[] {
     const maxItems = 4;
-    
+
     return movies.slice(0, maxItems).map((movie, index) => {
-      const baseTitle = movie.title || `${type === 'movie' ? 'Película' : 'Serie'} ${index + 1}`;
-      
+      // movie.title puede ser string o un objeto { value }
+      const rawTitle =
+        typeof movie?.title === 'string'
+          ? movie.title
+          : movie?.title?.value || movie?.name || '';
+
+      const baseTitle =
+        rawTitle || `${type === 'movie' ? 'Película' : 'Serie'} ${index + 1}`;
+
       return {
-        id: movie.id || `${type}_${index}`,
-        title: { 
-          value: type === 'series' ? `${baseTitle} (Serie)` : baseTitle 
+        id: movie.id || movie.tmdbId || `${type}_${index}`,
+        title: {
+          value: type === 'series' ? `${baseTitle} (Serie)` : baseTitle,
         },
-        category: { 
-          value: type === 'series' ? 'Series,Drama' : 'Cine,Drama' 
+        category: {
+          value:
+            type === 'series'
+              ? movie.category || 'Series,Drama'
+              : movie.category || 'Cine,Drama',
         },
         icon: this.getValidImageUrl(movie, index),
-        starRating: this.normalizeRating(movie.rating)
+        starRating: this.normalizeRating(movie.rating || movie.starRating),
       };
     });
   }
@@ -149,25 +190,25 @@ export class RightSidebarComponent implements OnInit {
    */
   private normalizeRating(rating: any): string {
     if (!rating) return '6.0';
-    
+
     if (typeof rating === 'number') {
       return rating.toFixed(1);
     }
-    
+
     if (typeof rating === 'string') {
       // Si ya tiene formato "/10", extraer solo el número
       if (rating.includes('/10')) {
         const parts = rating.split('/10');
         return parts[0] || '6.0';
       }
-      
+
       // Si es string número
       const numRating = parseFloat(rating);
       if (!isNaN(numRating)) {
         return numRating.toFixed(1);
       }
     }
-    
+
     return '6.0';
   }
 
@@ -187,14 +228,14 @@ export class RightSidebarComponent implements OnInit {
       'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiNkYzI2MjYiLz48cGF0aCBkPSJNNjAgNzBMMTAwIDk1TDYwIDEyMFY3MFpNNzUgNDBIODVWNjBINzVWNDBaTTc1IDE0MEg4NVYxNjBINzVWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=',
       'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiMzOTkzZGQiLz48cGF0aCBkPSJNNTAgNzBMMTAwIDk1TDUwIDEyMFY3MFpNNzAgNDBIODBWNjBINzBWNDBaTTcwIDE0MEg4MFYxNjBINzBWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=',
       'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiNmNTk1MjEiLz48cGF0aCBkPSJNNTUgNzBMMTAzIDk1TDU1IDEyMFY3MFpNNzIgNDBIODJWNjBINzJWNDBaTTcyIDE0MEg4MlYxNjBINzJWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=',
-      'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiM4YjVjZjYiLz48cGF0aCBkPSJNNTggNzBMMTAxIDk1TDU4IDEyMFY3MFpNNzQgNDBIODRWNjBINzRWNDBaTTc0IDE0MEg4NFYxNjBINzRWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo='
+      'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiM4YjVjZjYiLz48cGF0aCBkPSJNNTggNzBMMTAxIDk1TDU4IDEyMFY3MFpNNzQgNDBIODRWNjBINzRWNDBaTTc0IDE0MEg4NFYxNjBINzRWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=',
     ];
-    
+
     // Si hay poster válido de TMDb o similar, intentar usarlo
     if (movie.poster && this.isValidImageUrl(movie.poster)) {
       return movie.poster;
     }
-    
+
     // Usar imagen estática basada en el índice
     return staticImages[index % staticImages.length];
   }
@@ -204,14 +245,16 @@ export class RightSidebarComponent implements OnInit {
    */
   private isValidImageUrl(url: string): boolean {
     if (!url) return false;
-    
+
     // Evitar placeholders conocidos
-    if (url.includes('placeholder.com') || 
-        url.includes('via.placeholder') || 
-        url.includes('placehold')) {
+    if (
+      url.includes('placeholder.com') ||
+      url.includes('via.placeholder') ||
+      url.includes('placehold')
+    ) {
       return false;
     }
-    
+
     // Verificar que es una URL válida
     try {
       new URL(url);
@@ -229,10 +272,24 @@ export class RightSidebarComponent implements OnInit {
    * Navegar a detalles de una película/serie
    */
   public navigateTo(item: ISidebarItem): void {
-    console.log('🔗 RIGHT-SIDEBAR - Navigating to details:', item.title.value);
-    
-    const slug = item.title.value.replace(/\s/g, '-').toLowerCase();
-    this.router.navigate(['programacion-tv/detalles', slug]);
+    const rawTitle = item?.title?.value || '';
+    console.log('🔗 RIGHT-SIDEBAR - Navigating to details:', rawTitle);
+
+    // Obtener slug usando slugify para eliminar caracteres especiales
+    const slug = slugify(rawTitle);
+
+    // Heurística simple para distinguir película vs serie/programa:
+    // - categoría que empieza por 'Cine' -> película
+    // - o si tiene poster/tmdbId -> película
+    const cat = item?.category?.value || '';
+    const looksLikeMovie =
+      cat.startsWith('Cine') || !!(item as any).tmdbId || !!(item as any).icon;
+
+    if (looksLikeMovie) {
+      this.router.navigate(['/peliculas', slug]);
+    } else {
+      this.router.navigate(['/programas', slug]);
+    }
   }
 
   /**
@@ -240,10 +297,16 @@ export class RightSidebarComponent implements OnInit {
    */
   public navigateTo2(type: string): void {
     console.log('🔗 RIGHT-SIDEBAR - Navigating to list:', type);
-    
+
     if (type === 'movie') {
+      // marcar estado en el servicio para que ListaDestacadas cargue películas
+      try {
+        this.homeDataService.getCurrentState();
+      } catch {}
+      this.tvGuide.setIsMovies();
       this.router.navigate(['programacion-tv/peliculas']);
     } else if (type === 'serie') {
+      this.tvGuide.setIsSeries();
       this.router.navigate(['programacion-tv/series']);
     }
   }
@@ -260,7 +323,7 @@ export class RightSidebarComponent implements OnInit {
     if (movie.icon && !movie.icon.includes('picons_dobleM')) {
       return movie.icon; // Es un poster real, no un logo de canal
     }
-    
+
     // Si no hay poster válido, usar imagen por defecto
     return this.getDefaultPosterUrl();
   }
@@ -276,22 +339,36 @@ export class RightSidebarComponent implements OnInit {
    * Obtiene la categoría de una película
    */
   public getMovieCategory(movie: ISidebarItem): string {
-    if (movie.category?.value) {
-      const parts = movie.category.value.split(',');
-      return parts[1]?.trim() || parts[0]?.trim() || 'Drama';
-    }
-    return 'Drama';
+    if (!movie?.category?.value) return 'Drama';
+
+    const val: any = movie.category.value;
+    let raw = '';
+    if (typeof val === 'string') raw = val;
+    else if (Array.isArray(val)) raw = val.join(',');
+    else if (val && typeof val === 'object')
+      raw = val.value || val.name || Object.values(val).join(',');
+    else raw = String(val || '');
+
+    const parts = raw.split(',');
+    return parts[1]?.trim() || parts[0]?.trim() || 'Drama';
   }
 
   /**
    * Obtiene la categoría de una serie
    */
   public getSerieCategory(serie: ISidebarItem): string {
-    if (serie.category?.value) {
-      const parts = serie.category.value.split(',');
-      return parts[1]?.trim() || parts[0]?.trim() || 'Drama';
-    }
-    return 'Drama';
+    if (!serie?.category?.value) return 'Drama';
+
+    const val: any = serie.category.value;
+    let raw = '';
+    if (typeof val === 'string') raw = val;
+    else if (Array.isArray(val)) raw = val.join(',');
+    else if (val && typeof val === 'object')
+      raw = val.value || val.name || Object.values(val).join(',');
+    else raw = String(val || '');
+
+    const parts = raw.split(',');
+    return parts[1]?.trim() || parts[0]?.trim() || 'Drama';
   }
 
   /**
@@ -299,15 +376,15 @@ export class RightSidebarComponent implements OnInit {
    */
   public formatRating(rating: any): string {
     if (!rating) return 'N/A';
-    
+
     // Normalizar el rating primero
     const normalizedRating = this.normalizeRating(rating);
-    
+
     // Agregar /10 si no lo tiene
     if (!normalizedRating.includes('/10')) {
       return `${normalizedRating}/10`;
     }
-    
+
     return normalizedRating;
   }
 
@@ -334,14 +411,16 @@ export class RightSidebarComponent implements OnInit {
    * Track by function para películas
    */
   public trackByMovieId(index: number, movie: ISidebarItem): string {
-    return movie.id || movie.title?.value || index.toString();
+    const title = movie.title?.value || '';
+    return (movie.id || title || index.toString()).toString();
   }
 
   /**
    * Track by function para series
    */
   public trackBySerieId(index: number, serie: ISidebarItem): string {
-    return serie.id || serie.title?.value || index.toString();
+    const title = serie.title?.value || '';
+    return (serie.id || title || index.toString()).toString();
   }
 
   // ===============================================
@@ -356,18 +435,18 @@ export class RightSidebarComponent implements OnInit {
     console.table({
       'Popular Movies': this.popular_movies().length,
       'Popular Series': this.popular_series().length,
-      'Is Loading': this.isLoading()
+      'Is Loading': this.isLoading(),
     });
-    
+
     console.log('Movies data:', this.popular_movies());
     console.log('Series data:', this.popular_series());
-    
+
     // Debug del servicio también
     const serviceState = this.homeDataService.getCurrentState();
     console.log('Service state:', {
       featuredMovie: serviceState.featuredMovie?.title || 'none',
       popularMovies: serviceState.popularMovies.length,
-      isLoading: serviceState.isLoading
+      isLoading: serviceState.isLoading,
     });
   }
 
@@ -379,9 +458,9 @@ export class RightSidebarComponent implements OnInit {
     this.isLoading.set(true);
     this.popular_movies.set([]);
     this.popular_series.set([]);
-    
+
     // Trigger refresh desde el servicio
-    this.homeDataService.refreshData().subscribe(result => {
+    this.homeDataService.refreshData().subscribe((result) => {
       console.log('🔄 RIGHT-SIDEBAR - Refresh result:', result.success);
     });
   }
