@@ -3,14 +3,14 @@
  * Ubicación: src/app/pages/home/home.component.ts
  */
 
-import { 
-  Component, 
-  OnInit, 
-  ChangeDetectionStrategy, 
-  signal, 
+import {
+  Component,
+  OnInit,
+  ChangeDetectionStrategy,
+  signal,
   computed,
   inject,
-  DestroyRef
+  DestroyRef,
 } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { CommonModule } from '@angular/common';
@@ -23,10 +23,16 @@ import { NavBarComponent } from 'src/app/components/nav-bar/nav-bar.component';
 import { MetaService } from 'src/app/services/meta.service';
 import { HomeDataService } from 'src/app/services/features/home-data.service';
 import { CategoryFilterService } from 'src/app/services/program-list/category-filter.service';
-import { IFeaturedMovie, ITvProgram, ILogger, IDayChangedEvent } from 'src/app/interfaces';
+import {
+  IFeaturedMovie,
+  ITvProgram,
+  ILogger,
+  IDayChangedEvent,
+} from 'src/app/interfaces';
 import { ConsoleLoggerService } from 'src/app/services/core/logger.service';
 import { ProgramListComponent } from 'src/app/components/program-list/program-list.component';
 import { BannerComponent } from 'src/app/components/banner/banner.component';
+import { DeviceDetectorService } from 'src/app/services/device-detector.service';
 
 /**
  * HomeComponent - REFACTORIZADO CON PRINCIPIOS SOLID
@@ -37,19 +43,23 @@ import { BannerComponent } from 'src/app/components/banner/banner.component';
   styleUrls: ['./home.component.scss'],
   standalone: true,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CommonModule, NavBarComponent, ProgramListComponent, BannerComponent],
+  imports: [
+    CommonModule,
+    NavBarComponent,
+    ProgramListComponent,
+    BannerComponent,
+  ],
 })
 export class HomeComponent implements OnInit {
-  
   // ===============================================
   // DEPENDENCY INJECTION - DEPENDENCY INVERSION PRINCIPLE
   // ===============================================
-  
+  public readonly deviceDetector = inject(DeviceDetectorService);
+
   private readonly destroyRef = inject(DestroyRef);
   private readonly metaService = inject(MetaService);
   private readonly homeDataService = inject(HomeDataService);
   private readonly categoryFilterService = inject(CategoryFilterService);
-  private readonly router = inject(Router);
   private readonly logger = inject(ConsoleLoggerService);
 
   // ===============================================
@@ -60,31 +70,34 @@ export class HomeComponent implements OnInit {
   public readonly programs = signal<ITvProgram[]>([]);
   public readonly featuredMovie = signal<IFeaturedMovie | null>(null);
   public readonly popularMovies = signal<IFeaturedMovie[]>([]);
-  
+
   // Estados de UI
   public readonly isLoading = signal<boolean>(true);
   public readonly error = signal<string | null>(null);
-  
+
   // Categorías gestionadas por el servicio dedicado
-  public readonly selectedCategory = computed(() => this.categoryFilterService.getSelectedCategory());
+  public readonly selectedCategory = computed(() =>
+    this.categoryFilterService.getSelectedCategory()
+  );
 
   // ===============================================
   // COMPUTED PROPERTIES - DERIVED STATE
   // ===============================================
 
   public readonly hasData = computed(() => this.programs().length > 0);
-  public readonly hasFeaturedMovie = computed(() => this.featuredMovie() !== null);
+  public readonly hasFeaturedMovie = computed(
+    () => this.featuredMovie() !== null
+  );
   public readonly hasError = computed(() => this.error() !== null);
-  
+
   // Estado combinado para la UI
   public readonly uiState = computed(() => ({
     isLoading: this.isLoading(),
-    hasData: this.hasData(),
     hasFeaturedMovie: this.hasFeaturedMovie(),
     hasError: this.hasError(),
     showContent: !this.isLoading() && this.hasData() && !this.hasError(),
     showError: !this.isLoading() && this.hasError(),
-    showEmpty: !this.isLoading() && !this.hasData() && !this.hasError()
+    showEmpty: !this.isLoading() && !this.hasData() && !this.hasError(),
   }));
 
   // ===============================================
@@ -93,15 +106,16 @@ export class HomeComponent implements OnInit {
 
   ngOnInit(): void {
     this.logger.info('HomeComponent initializing');
-   
+    // NUEVO: Asegurar que DeviceDetector está inicializado
+
     this.setupMetaTags();
     this.initializeDataStreams();
     this.initializeData();
-    
+
     // AÑADIDO: Exponer métodos de debug en desarrollo
     if (!this.isProduction()) {
       this.exposeDebugMethods();
-      
+
       // Verificar consistencia después de la inicialización
       setTimeout(() => {
         this.checkDataConsistency();
@@ -119,8 +133,9 @@ export class HomeComponent implements OnInit {
   private setupMetaTags(): void {
     this.metaService.setMetaTags({
       title: 'Guía TV - Programación de Televisión Actual',
-      description: 'Descubre qué ver en TV hoy. Guía completa de programación televisiva con horarios actualizados.',
-      canonicalUrl: '/'
+      description:
+        'Descubre qué ver en TV hoy. Guía completa de programación televisiva con horarios actualizados.',
+      canonicalUrl: '/',
     });
   }
 
@@ -133,10 +148,10 @@ export class HomeComponent implements OnInit {
     // Stream de programas
     this.homeDataService.programs$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(programs => {
+      .subscribe((programs) => {
         this.logger.debug(`Programs updated: ${programs.length} items`);
         this.programs.set(programs);
-        
+
         // Actualizar programas en el servicio de categorías
         this.categoryFilterService.updatePrograms(programs);
       });
@@ -144,7 +159,7 @@ export class HomeComponent implements OnInit {
     // Stream de película destacada
     this.homeDataService.featuredMovie$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(movie => {
+      .subscribe((movie) => {
         this.logger.debug(`Featured movie updated: ${movie?.title || 'none'}`);
         this.logger.debug(`Featured movie details:`, movie);
         this.featuredMovie.set(movie);
@@ -153,7 +168,7 @@ export class HomeComponent implements OnInit {
     // Stream de películas populares
     this.homeDataService.popularMovies$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(movies => {
+      .subscribe((movies) => {
         this.logger.debug(`Popular movies updated: ${movies.length} items`);
         this.popularMovies.set(movies);
       });
@@ -161,7 +176,7 @@ export class HomeComponent implements OnInit {
     // Stream de estado de carga
     this.homeDataService.loading$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(loading => {
+      .subscribe((loading) => {
         this.logger.debug(`Loading state: ${loading}`);
         this.isLoading.set(loading);
       });
@@ -169,7 +184,7 @@ export class HomeComponent implements OnInit {
     // Stream de errores
     this.homeDataService.error$
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(error => {
+      .subscribe((error) => {
         this.logger.debug(`Error state: ${error || 'none'}`);
         this.error.set(error);
       });
@@ -181,9 +196,10 @@ export class HomeComponent implements OnInit {
   private initializeData(): void {
     this.logger.info('Starting data initialization');
 
-    this.homeDataService.initializeData()
+    this.homeDataService
+      .initializeData()
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(result => {
+      .subscribe((result) => {
         if (result.success) {
           this.logger.info('Home data initialization successful');
         } else {
@@ -201,19 +217,25 @@ export class HomeComponent implements OnInit {
    */
   public onDayChanged(event: IDayChangedEvent): void {
     const { dayIndex, dayInfo } = event;
-    
-    this.logger.info(`Day changed: index ${dayIndex}, day: ${dayInfo.diaSemana} ${dayInfo.diaNumero}`);
-    
+
+    this.logger.info(
+      `Day changed: index ${dayIndex}, day: ${dayInfo.diaSemana} ${dayInfo.diaNumero}`
+    );
+
     // Actualizar el título de la página según el día seleccionado
     if (dayIndex === 0) {
       // Hoy
       this.updatePageTitle('Guía TV - Programación de Hoy');
     } else if (dayIndex === 1) {
       // Mañana
-      this.updatePageTitle(`Guía TV - Programación de Mañana (${dayInfo.diaSemana} ${dayInfo.diaNumero})`);
+      this.updatePageTitle(
+        `Guía TV - Programación de Mañana (${dayInfo.diaSemana} ${dayInfo.diaNumero})`
+      );
     } else {
       // Pasado mañana
-      this.updatePageTitle(`Guía TV - Programación ${dayInfo.diaSemana} ${dayInfo.diaNumero}`);
+      this.updatePageTitle(
+        `Guía TV - Programación ${dayInfo.diaSemana} ${dayInfo.diaNumero}`
+      );
     }
   }
 
@@ -232,9 +254,10 @@ export class HomeComponent implements OnInit {
   public onRetry(): void {
     this.logger.info('Retry requested by user');
     this.error.set(null);
-    this.homeDataService.refreshData()
+    this.homeDataService
+      .refreshData()
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(result => {
+      .subscribe((result) => {
         if (!result.success) {
           this.logger.error(`Retry failed: ${result}`);
         }
@@ -246,9 +269,10 @@ export class HomeComponent implements OnInit {
    */
   public onRefresh(): void {
     this.logger.info('Manual refresh requested');
-    this.homeDataService.refreshData()
+    this.homeDataService
+      .refreshData()
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(result => {
+      .subscribe((result) => {
         if (result.success) {
           this.logger.info('Manual refresh completed');
         } else {
@@ -267,9 +291,9 @@ export class HomeComponent implements OnInit {
   public formatTime(dateString: string): string {
     try {
       const date = new Date(dateString);
-      return date.toLocaleTimeString('es-ES', { 
-        hour: '2-digit', 
-        minute: '2-digit' 
+      return date.toLocaleTimeString('es-ES', {
+        hour: '2-digit',
+        minute: '2-digit',
       });
     } catch (error) {
       this.logger.warn(`Invalid date string: ${dateString}`);
@@ -282,7 +306,7 @@ export class HomeComponent implements OnInit {
    */
   public formatRating(rating: number | string | null): string {
     if (!rating) return 'N/A';
-    
+
     // Si ya es una string con formato (ej: "6.0/10"), devolverla tal como está
     if (typeof rating === 'string') {
       // Si contiene "/10" o "/", devolverlo tal como está
@@ -296,12 +320,12 @@ export class HomeComponent implements OnInit {
       }
       return rating; // Devolver la string original si no se puede procesar
     }
-    
+
     // Si es número, formatear normalmente
     if (typeof rating === 'number') {
       return `${rating.toFixed(1)}`;
     }
-    
+
     return 'N/A';
   }
 
@@ -338,20 +362,24 @@ export class HomeComponent implements OnInit {
       icon: featured.poster || 'assets/images/default-movie-poster.svg',
       poster: featured.poster || 'assets/images/default-movie-poster.svg',
       start: featured.startTime || new Date().toISOString(),
-      stop: featured.endTime || new Date(Date.now() + 2 * 60 * 60 * 1000).toISOString(),
+      stop:
+        featured.endTime ||
+        new Date(Date.now() + 2 * 60 * 60 * 1000).toISOString(),
       startTime: featured.startTime,
       endTime: featured.endTime,
       desc: {
-        details: featured.description || 'Película destacada de la programación actual.',
+        details:
+          featured.description ||
+          'Película destacada de la programación actual.',
         year: featured.releaseDate,
-        rate: 'TP'
+        rate: 'TP',
       },
       description: featured.description,
       year: featured.releaseDate,
       rating: featured.rating,
       starRating: featured.rating || '7.0',
       category: featured.category || 'Cine',
-      id: featured.id
+      id: featured.id,
     };
   }
 
@@ -359,7 +387,10 @@ export class HomeComponent implements OnInit {
    * Obtiene la URL del poster con fallback mejorado
    */
   public getPosterUrl(movie: IFeaturedMovie): string {
-    return movie.poster || 'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiMzOTkzZGQiLz48cGF0aCBkPSJNNTAgNzBMMTAwIDk1TDUwIDEyMFY3MFpNNzAgNDBIODBWNjBINzBWNDBaTTcwIDE0MEg4MFYxNjBINzBWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=';
+    return (
+      movie.poster ||
+      'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiMzOTkzZGQiLz48cGF0aCBkPSJNNTAgNzBMMTAwIDk1TDUwIDEyMFY3MFpNNzAgNDBIODBWNjBINzBWNDBaTTcwIDE0MEg4MFYxNjBINzBWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo='
+    );
   }
 
   /**
@@ -367,8 +398,8 @@ export class HomeComponent implements OnInit {
    */
   public onPosterError(event: Event): void {
     const img = event.target as HTMLImageElement;
-    img.src = 'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiMzOTkzZGQiLz48cGF0aCBkPSJNNTAgNzBMMTAwIDk1TDUwIDEyMFY3MFpNNzAgNDBIODBWNjBINzBWNDBaTTcwIDE0MEg4MFYxNjBINzBWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=';
-
+    img.src =
+      'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTUwIiBoZWlnaHQ9IjIwMCIgdmlld0JveD0iMCAwIDE1MCAyMDAiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+PHJlY3Qgd2lkdGg9IjE1MCIgaGVpZ2h0PSIyMDAiIGZpbGw9IiMzOTkzZGQiLz48cGF0aCBkPSJNNTAgNzBMMTAwIDk1TDUwIDEyMFY3MFpNNzAgNDBIODBWNjBINzBWNDBaTTcwIDE0MEg4MFYxNjBINzBWMTQwWiIgZmlsbD0iI2ZmZmZmZiIvPjwvc3ZnPgo=';
   }
 
   /**
@@ -388,8 +419,9 @@ export class HomeComponent implements OnInit {
   private updatePageTitle(title: string): void {
     this.metaService.setMetaTags({
       title,
-      description: 'Descubre qué ver en TV. Guía completa de programación televisiva con horarios actualizados.',
-      canonicalUrl: '/'
+      description:
+        'Descubre qué ver en TV. Guía completa de programación televisiva con horarios actualizados.',
+      canonicalUrl: '/',
     });
   }
 
@@ -412,12 +444,14 @@ export class HomeComponent implements OnInit {
       this.logger.debug('=== HOME COMPONENT STATE ===');
       this.logger.debug(`UI State:`, this.uiState());
       this.logger.debug(`Programs: ${this.programs().length}`);
-      this.logger.debug(`Featured Movie: ${this.featuredMovie()?.title || 'none'}`);
+      this.logger.debug(
+        `Featured Movie: ${this.featuredMovie()?.title || 'none'}`
+      );
       this.logger.debug(`Popular Movies: ${this.popularMovies().length}`);
-      
+
       // Debug del servicio de datos
       this.homeDataService.debugState();
-      
+
       this.logger.debug('=== END COMPONENT STATE ===');
     }
   }
@@ -427,16 +461,19 @@ export class HomeComponent implements OnInit {
    */
   public forceSyncData(): void {
     this.logger.info('🔄 FORCE SYNC - Forcing data synchronization');
-    
+
     const currentState = this.homeDataService.getCurrentState();
-    
+
     // Si hay datos en el servicio pero no en el componente, forzar sincronización
-    if (currentState.programListData.length > 0 && this.programs().length === 0) {
+    if (
+      currentState.programListData.length > 0 &&
+      this.programs().length === 0
+    ) {
       this.logger.warn('⚠️ FORCE SYNC - Data mismatch detected, forcing sync');
-      
+
       // Forzar actualización de datos en ProgramList
       this.homeDataService.updateProgramListData(currentState.programListData);
-      
+
       // Forzar change detection
       setTimeout(() => {
         this.logger.info('🔄 FORCE SYNC - Forcing change detection');
@@ -452,9 +489,9 @@ export class HomeComponent implements OnInit {
     const componentState = {
       programs: this.programs().length,
       featuredMovie: this.featuredMovie()?.title || 'none',
-      popularMovies: this.popularMovies().length
+      popularMovies: this.popularMovies().length,
     };
-    
+
     console.log('🔍 DATA CONSISTENCY CHECK:');
     console.log('Service State:', {
       programs: serviceState.programs.length,
@@ -462,19 +499,26 @@ export class HomeComponent implements OnInit {
       featuredMovie: serviceState.featuredMovie?.title || 'none',
       popularMovies: serviceState.popularMovies.length,
       isLoading: serviceState.isLoading,
-      hasData: serviceState.hasData
+      hasData: serviceState.hasData,
     });
-    
+
     console.log('Component State:', componentState);
-    
+
     // Detectar inconsistencias
-    if (serviceState.programListData.length > 0 && this.programs().length === 0) {
-      console.warn('🚨 INCONSISTENCY: Service has ProgramList data but component has no programs');
+    if (
+      serviceState.programListData.length > 0 &&
+      this.programs().length === 0
+    ) {
+      console.warn(
+        '🚨 INCONSISTENCY: Service has ProgramList data but component has no programs'
+      );
       this.forceSyncData();
     }
-    
+
     if (serviceState.featuredMovie && !this.featuredMovie()) {
-      console.warn('🚨 INCONSISTENCY: Service has featured movie but component does not');
+      console.warn(
+        '🚨 INCONSISTENCY: Service has featured movie but component does not'
+      );
     }
   }
 
@@ -493,19 +537,23 @@ export class HomeComponent implements OnInit {
           programs: this.programs().length,
           featuredMovie: this.featuredMovie()?.title,
           popularMovies: this.popularMovies().length,
-          uiState: this.uiState()
+          uiState: this.uiState(),
         }),
-        testFormatRating: (value: any) => this.formatRating(value)
+        testFormatRating: (value: any) => this.formatRating(value),
       };
-      
+
       console.log('🛠️ DEBUG METHODS EXPOSED:');
       console.log('- homeComponentDebug.state() - Ver estado completo');
       console.log('- homeComponentDebug.sync() - Forzar sincronización');
       console.log('- homeComponentDebug.check() - Verificar consistencia');
       console.log('- homeComponentDebug.refresh() - Refrescar datos');
       console.log('- homeComponentDebug.serviceState() - Estado del servicio');
-      console.log('- homeComponentDebug.componentState() - Estado del componente');
-      console.log('- homeComponentDebug.testFormatRating(value) - Probar formatRating');
+      console.log(
+        '- homeComponentDebug.componentState() - Estado del componente'
+      );
+      console.log(
+        '- homeComponentDebug.testFormatRating(value) - Probar formatRating'
+      );
     }
   }
 
@@ -519,24 +567,32 @@ export class HomeComponent implements OnInit {
    * ACTUALIZADO: Ahora maneja selección múltiple
    */
   public onCategorySelected(categories: string[]): void {
-    this.logger.info(`HomeComponent: Categorías seleccionadas: ${categories.join(', ')}`);
-    
+    this.logger.info(
+      `HomeComponent: Categorías seleccionadas: ${categories.join(', ')}`
+    );
+
     // Si no hay categorías seleccionadas, limpiar filtro
     if (categories.length === 0) {
       this.categoryFilterService.clearCategoryFilter();
       return;
     }
-    
+
     // Por ahora, el CategoryFilterService maneja una sola categoría
     // Usamos la primera categoría seleccionada como principal
     // TODO: Actualizar CategoryFilterService para soportar múltiples categorías
     const primaryCategory = categories[0];
     this.categoryFilterService.selectCategory(primaryCategory);
-    
+
     // Log adicional para múltiples categorías
     if (categories.length > 1) {
-      this.logger.info(`HomeComponent: Múltiples categorías seleccionadas, usando como principal: ${primaryCategory}`);
-      this.logger.debug(`HomeComponent: Categorías adicionales: ${categories.slice(1).join(', ')}`);
+      this.logger.info(
+        `HomeComponent: Múltiples categorías seleccionadas, usando como principal: ${primaryCategory}`
+      );
+      this.logger.debug(
+        `HomeComponent: Categorías adicionales: ${categories
+          .slice(1)
+          .join(', ')}`
+      );
     }
   }
 
@@ -557,7 +613,7 @@ export class HomeComponent implements OnInit {
   public getCategoryProgramCount(): number {
     const category = this.selectedCategory();
     if (!category) return 0;
-    
+
     const stats = this.categoryFilterService.getCategoryStats(category);
     return stats.totalPrograms;
   }
@@ -570,7 +626,7 @@ export class HomeComponent implements OnInit {
   public getCategoryChannelCount(): number {
     const category = this.selectedCategory();
     if (!category) return 0;
-    
+
     const stats = this.categoryFilterService.getCategoryStats(category);
     return stats.channelsCount;
   }
@@ -583,7 +639,7 @@ export class HomeComponent implements OnInit {
   public getCurrentCategoryPrograms(): number {
     const category = this.selectedCategory();
     if (!category) return 0;
-    
+
     const stats = this.categoryFilterService.getCategoryStats(category);
     return stats.currentlyAiring;
   }
@@ -596,7 +652,7 @@ export class HomeComponent implements OnInit {
   public getNextCategoryProgram(): string {
     const category = this.selectedCategory();
     if (!category) return '0';
-    
+
     const stats = this.categoryFilterService.getCategoryStats(category);
     return stats.nextProgramTime;
   }
