@@ -1,12 +1,14 @@
 export const environment = {
   production: true,
   firebaseConfig: {
-    apiKey: "AIzaSyCK_jTongKez-uHqn0Inoh19DicWNPAv_o",
-    authDomain: "guia-tv-8fe3c.firebaseapp.com",
-    projectId: "guia-tv-8fe3c",
-    storageBucket: "guia-tv-8fe3c.appspot.com",
-    messagingSenderId: "533857766582",
-    appId: "1:533857766582:web:4dfb40ca1ac9bb39346a17",
-    measurementId: "G-P6Z3BY2DVW"
-  }
+    apiKey: 'AIzaSyCK_jTongKez-uHqn0Inoh19DicWNPAv_o',
+    authDomain: 'guia-tv-8fe3c.firebaseapp.com',
+    projectId: 'guia-tv-8fe3c',
+    storageBucket: 'guia-tv-8fe3c.appspot.com',
+    messagingSenderId: '533857766582',
+    appId: '1:533857766582:web:4dfb40ca1ac9bb39346a17',
+    measurementId: 'G-P6Z3BY2DVW',
+  },
+  // REEMPLAZA este valor por tu dominio en producción
+  SITE_URL: 'https://www.guia-tv.com',
 };
